requirements.txt
customtkinter
yfinance
pandas
pandas_ta
numpy
matplotlib
mplfinance
google-generativeai
requests
python-dotenv
qrcode
Pillow
pytest
pytest-mock
python-dateutil
